Major update to instructions
I tried to make it so that immediate could be handled. It works for the most part, but it seems that LLVM handles immediates in such a way that it only makes sense to ever have an immediate on the left-side operand. LLVM actually complains if you have it on the right-side. We can look more into that